feat(helper): chapter_01 for mentors
In this chapter, nothing too complicated for now.
Just a new note on the chapters more generally: chapters will require your trainee(s) to be more and more autonomous.

Here, some set up in this chapter and a few learning resources.

Thoses resources are essentially preparing your trainee to be confortable with the GCP Console.

Be patient for each chapter, it will take the time it needs to complete: no deadline

The D-Day:
- Just be present, answer the questions
- Explain again the context. Emphasize it is a realistic use case.
- Explain shortly from the Project Architecture image how the project will be decoupled.
- Check everything is set up
- Check his account on A Cloud Guru is activated: acloudguru.devoteamgcloud.com ✨
- Reassure your trainee(s) to go at his own path. No deadline. The main point is to be confortable with the notions seen.
- As a bonus, go through the Google Cloud Developer Cheat Sheet: https://googlecloudcheatsheet.withgoogle.com/
- Have fun !!!