feat(helper): chapter_03 for mentors
In this chapter, a lot to learn, less to do.
Particurlarly in this chapter, be cautious with your trainee(s).

Some have never coded in Python.

Here the main point is to complete an already existing code.
The trainee(s) will have to handle errors. (And of course, understand the code!)

Explain again what `assert` is used for.
Note the `[MENTOR #1]` where the `assert` is handled by the `datetime` library.

In this chapter, emphasize on the fact it is easier sometimes to test locally on mock data (in the `__main__` of the code)
The trainee(s) will again have to create the Cloud function first from the Console, then with the CLI then with Terraform.

Explain again Terraform if some questions remain.

For more advanced trainee(s), ask them to read PEP20: https://peps.python.org/pep-0020/
And ask them to follow PEP8 (Python Style Guide): https://peps.python.org/pep-0008/

When the chapter is considered as finished, you can remove the simulating return as shonw with the `[MENTOR #2]`

The D-Day:
- Just be present, answer the questions.
- Explain again the context.
- Try to get from the traine(s) the theorical points they learn. Enrich this learning.
- Do not hesitate to enrich with your own experience.
- It is of if you do not have all the answer at the time it is asked.
- Have fun !!!