feat(helper): chapter_04 for mentors
This chapter is hard ! And this chapter is long !

First, a lot of new resources to create with Terraform.
It will be ask to deploy the queries and schemas files.

Note the `[MENTOR #1]` where a `for_each` is used in the Terraform.
You can introduce this new concept to your trainee(s) before or during the code reviews.
This way, you can upload all the files relative to BigQuery in one resource.

Important, you can explain to your trainee(s) what the `magasin-cie-utlils` bucket is useful for.
And why we need a second bucket only to store those files.
The main reasons is to decouple usages:
- the `magasin-cie-landing` is here as a landplace to the files from the client.
  It also handles the lifecyle of those files as well as organizing them in folder for easy idendification and debug.
- the `magasin-cie-utlils` bucket will serve as a landplace to make your pipeline work.
  It will store all the files needed to be retrieve during runtime.
In short, the first bucket is for data, the second is for engineering.

Check the understanding of PubSub too and explain briefly what Cloud Workflows is.
Cloud Workflows is the subject of the next chapter so you can stay high level.

Then to deploy BigQuery datasets and Tables in Terraform, you can also use `for_each`
to anticipate future tables as written in the `[MENTOR #2]`

You can also introduce the importance of naming this correctily and in a descriptive way.
For instance for the Cloud Functions in Terrafom, ([MENTOR #3]), the name of the resources in the deployed last chapter
was too generic. So to create a second Cloud Functions I need to be more intelligent and I had to refactor my code.
Example:
  `source` is a bad name `source_<specific_name>` is better
  and it avoids having to go over all the code again because of a bad naming

For the Cloud Function itself and the Cloud Workflows, there is an error with default values :)
Your trainee(s) will probably truggle to find `magasin-cie-utils` does not exists but is instead `magasin_cie_utils`
You can find those mispelling in the `env.yaml` of the Cloud Functions and in the assignements of the Cloud Workflows ([MENTOR 4])

And more precisely in the Cloud Function code, your trainee(s) will have to interact with a loooot of GCP services with Python.
Important: It will take the time it takes!
Important: It will take the time it takes!

Be there to explain this. is. really. hard. part!
So it is ok to get errors and errors and feel like it is impossible or hard.

Be patient and invite your trainee(s) to ask other teammates for help if they need :)

On your side, just be carefull of infinite loop that you do not want.
For instance, if your Cloud Functions recreate the files in the `input/` path then you first Cloud Function will be triggered again, and to the infinite.

Again, take your time as it will be nice to ask for PEP8 conventions in the code.

When everything is done, you can test the full pipeline for the `store` file activating the PubSub.  `[MENTOR #5]`
Before everything is well check, invite your trainee(s) to send manual messages to PubSub to avoid infinite Loop (and interact with PubSub too)

Note also the `[MENTOR #6]`. Explain why it is in `WRITE_TRUNCATE` mode for `write_disposition`.
You can also explain the `create_disposition` modes which are not present in this code.

The D-Day:
- Just be present, answer the questions.
- Explain again the context.
- Try to get from the traine(s) the theorical points they learn. Enrich this learning.
- Do not hesitate to enrich with your own experience.
- It is of if you do not have all the answer at the time it is asked.
- Have fun !!!