feat(helper): chapter_05 for mentors
Nothing highly technical here. But it can be highly difficult to handle as this
chapter is here to transcribe a FUNCTIONAL use case to queries and tables.

Here we start from the `Interface Contracts & Mappings` Google Sheet and the trainee(s)
will have to transcribe this understandable way of thinking data into a technical answer.
In short, create a Workflow with queries In BigQuery.

First, do not hesitate to explain what is a Schema table for BigQuery.
Then try to enrich with your own experience as it is highly functional.

First, explain again what is the difference of delta or full file sents:
- full: the data is sent with every element from the source. The file contains the full view of the source.
- delta: the data is sent only with changes from last sent.

This is why we need to use a `MERGE` statement here.
It will only insert new data and update existing data (given primary keys for identification).
So every element inserted will persist over time and will not be deleted until a manual deletion.

The `MERGE` statement is helpful but we have to be careful with the source data.
It requires not to have duplicated PK in it! Otherwise, errors will be raised next times.
Let's take an example.

Imagine we want to insert from A `[{'PK': 'abc', 'value': 123}, {'PK': 'abc', 'value': 456}]` in an empty table B with a `MERGE` statement on `PK`

`PK='abc'` is not found in B so the two lines will be inserted.
Then we want to update this data from A' `[{'PK': 'abc', 'value': 789}]`

`PK='abc'` will match two rows in B so it will raise an error and the table is then inaccessible anymore.

This is why we use `QUALIFY ROW_NUMBER() OVER(...) = 1` in `[MENTOR #1]` to deduplicates the data to avoid future errors.
Check the article to know more about this.

Then, imagine we integrated a file from March 1st, and for some reasons, the source sends us a file from January (late file data for instance)
You can add a condition in the `MATCH` condition of the `MERGE` as shown with `[MENTOR #2]` to only update the data based on freshness comparison.
If you do not do this, you will have to reload all the files one by one from January until today.

The D-Day:
- Just be present, answer the questions.
- Explain again the context.
- Try to get from the traine(s) the theorical points they learn. Enrich this learning.
- Do not hesitate to enrich with your own experience.
- It is of if you do not have all the answer at the time it is asked.
- Have fun !!!