feat(helper): chapter_06 for mentors
In this last chapter on the Data Pipeline... It will be hard ! Really Hard !

Trainee(s) will be uncomfortable and this is the point.
They have to take as much time as it will need to complete.

It will be hard for a lot of reasons including:
- highly complex because they need to master all the services seen since.
- complicated as it will need some advanced knowledge
- it demands to be highly autonomous as the mission is not describing a lot (almost a "do the project")
- it requires to really understand data and functional ideas from the mappings
- some new stuff will be seen (views, DML, json, STRUCT, REPEATED, ...)
- ...

In short, a real-world scenario.
So take the time with your trainee(s) and encourage them to ask for some help with other teammates.

Ok this chapter is hard, but you still have to be demanding on the quality on the code produced.

First, it will probably need to rework the Cloud Function to dispatch to workflows to handle JSON files now
The Cloud Function has to load CSV and JSON now.

Nothing too hard on the Terraform side in this chapter.

The longest part will be to create ALL the schemas of the tables from the mapping.
Then invite your trainees to go to the GCP Console on BigQuery to iterate and test their query faster (and better)
before to integrate the queries into the iac.

`MERGE` queries will be a little bit harder to understand because of the difference of the fields `creation_time` and `update_time`
So take the time to explain why those two fields exists.

A lot of issues will be raised because of the RECORD REPEATED.
They will have to manage ARRAY, STRUCT, UNNEST, CROSS JOIN, ...
Take the time to redirect them to nice resources and to explain those concepts. Better with your own experience.

The hardest part will be the creation of the `staging/basket.sql` query.
They will probably not succeed.
If this is the case after 2 days of struggeling. Please, check by yourself the solution,
and explain to them the solution point by point, line by line.
But please, do not reveal the solution exists on this branch

Aggregated tables will be easier ;)
Just a point on the GDRP to enrich with your own experience and words.

The D-Day:
- Just be present, answer the questions.
- Explain again the context.
- Try to get from the traine(s) the theorical points they learn. Enrich this learning.
- Do not hesitate to enrich with your own experience.
- It is of if you do not have all the answer at the time it is asked.
- Have fun !!!